add chain info item
