+++ nekoyume/Assets/_Scripts/UI/MainCanvas.cs
@@ -2,12 +2,7 @@ using System;
 using System.Collections;
 using System.Collections.Generic;
 using System.Linq;
-using System.Threading.Tasks;
 using Nekoyume.EnumType;
-using Nekoyume.Helper;
-using Nekoyume.L10n;
-using Nekoyume.Game;
-using Nekoyume.Model.Mail;
 using Nekoyume.Pattern;
 using Nekoyume.UI.Module;
 using UnityEngine;
@@ -470,6 +465,8 @@ namespace Nekoyume.UI
             _secondWidgets.Add(Widget.Create<CustomCraftResultPopup>());
             yield return null;
             _secondWidgets.Add(Widget.Create<CraftResultPopup>());
+            yield return null;
+            _secondWidgets.Add(Widget.Create<ChainInfoPopup>());
             yield return null;
 
             // tooltip
