+++ nekoyume/Assets/_Scripts/UI/Widget/ChainInfoItem.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 34682b77ffd70c04c91b001e7defdb0a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
